Add CookieBox, CookieSwap and Baked Bazaar adapters
